chore(server): move all gateways to a dedicated module
