--- .vscode/extensions.json
@@ -1,7 +1,0 @@
-{
-  "recommendations": [
-    
-    "nrwl.angular-console",
-    "esbenp.prettier-vscode"
-  ]
-}
