+++ libs/server/gateway/README.md
@@ -1,0 +1,5 @@
+# libs/server/gateway
+
+This library is part of [`apps/server`](../../../apps/server) application.
+
+It contains all implementation of `gateway`s defined by the `libs/server/domain` layer.
