+++ libs/server/gateway/build.gradle.kts
@@ -17,46 +17,26 @@ val quarkusPlatformGroupId: String by project
 val quarkusPlatformArtifactId: String by project
 val quarkusPlatformVersion: String by project
 val ktlintVersion: String by project
+val atriumVersion: String by project
 val mockkVersion: String by project
 val quarkusMockkVersion: String by project
-val atriumVersion: String by project
-val cliktVersion: String by project
-val jacksonDatatypeJsr310Version: String by project
 
 dependencies {
-  implementation(kotlin("stdlib-jdk8"))
-  implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-
   implementation(enforcedPlatform("$quarkusPlatformGroupId:$quarkusPlatformArtifactId:$quarkusPlatformVersion"))
-  implementation("io.quarkus:quarkus-arc")
-  implementation("io.quarkus:quarkus-config-yaml")
   implementation("io.quarkus:quarkus-kotlin")
   implementation("io.quarkus:quarkus-mongodb-panache-kotlin")
-  implementation("io.quarkus:quarkus-resteasy-reactive")
-  implementation("io.quarkus:quarkus-resteasy-reactive-jackson")
-  implementation("io.quarkus:quarkus-security")
-  implementation("io.quarkus:quarkus-smallrye-context-propagation")
-  implementation("io.quarkus:quarkus-smallrye-openapi")
-  implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:$jacksonDatatypeJsr310Version")
-  implementation("com.github.ajalt.clikt:clikt:$cliktVersion")
 
   implementation(project(":libs:server:domain"))
-  implementation(project(":libs:server:gateway"))
   implementation(project(":libs:server:persistence"))
-  implementation(project(":libs:server:storage:azure"))
   implementation(project(":libs:server:storage:core"))
-  implementation(project(":libs:server:storage:gcs"))
-  implementation(project(":libs:server:storage:s3"))
 
   testImplementation("org.jetbrains.kotlinx:kotlinx-coroutines-test")
   testImplementation("ch.tutteli.atrium:atrium-fluent:$atriumVersion")
   testImplementation("io.mockk:mockk:$mockkVersion")
   testImplementation("io.quarkiverse.mockk:quarkus-junit5-mockk:$quarkusMockkVersion")
   testImplementation("io.quarkus:quarkus-junit5")
-  testImplementation("io.quarkus:quarkus-junit5-internal")
   testImplementation("io.quarkus:quarkus-jacoco")
   testImplementation("io.quarkus:quarkus-test-hibernate-reactive-panache")
-  testImplementation("io.rest-assured:rest-assured")
 }
 
 group = "org.nxcloudce.server"
@@ -67,14 +47,16 @@ java {
   targetCompatibility = JavaVersion.toVersion(javaVersion)
 }
 
+tasks.withType<Jar> {
+  archiveBaseName = "${project.group}.${project.name}"
+}
+
 tasks.withType<Test> {
   systemProperty("java.util.logging.manager", "org.jboss.logmanager.LogManager")
 }
 
 allOpen {
-  annotation("jakarta.ws.rs.Path")
   annotation("jakarta.enterprise.context.ApplicationScoped")
-  annotation("jakarta.persistence.Entity")
   annotation("io.quarkus.test.junit.QuarkusTest")
 }
