+++ libs/server/gateway/src/main/kotlin/org/nxcloudce/server/gateway/persistence/AccessTokenMapper.kt
@@ -1,0 +1,14 @@
+package org.nxcloudce.server.gateway.persistence
+
+import org.nxcloudce.server.domain.workspace.model.*
+import org.nxcloudce.server.persistence.entity.AccessTokenEntity
+
+fun AccessTokenEntity.toDomain(): AccessToken =
+  AccessToken {
+    id = AccessTokenId(this@toDomain.id.toString())
+    name = this@toDomain.name
+    publicId = AccessTokenPublicId(this@toDomain.publicId)
+    accessLevel = AccessLevel.from(this@toDomain.accessLevel)
+    workspaceId = WorkspaceId(this@toDomain.workspaceId.toString())
+    encodedValue = this@toDomain.encodedValue
+  }
