+++ libs/server/gateway/src/main/kotlin/org/nxcloudce/server/gateway/persistence/OrganizationMapper.kt
@@ -1,0 +1,10 @@
+package org.nxcloudce.server.gateway.persistence
+
+import org.nxcloudce.server.domain.organization.model.Organization
+import org.nxcloudce.server.domain.organization.model.OrganizationId
+import org.nxcloudce.server.domain.organization.usecase.CreateOrganizationRequest
+import org.nxcloudce.server.persistence.entity.OrganizationEntity
+
+fun OrganizationEntity.toDomain() = Organization(id = OrganizationId(id.toString()), name = name)
+
+fun CreateOrganizationRequest.toEntity() = OrganizationEntity(id = null, name = name)
