+++ settings.gradle.kts
@@ -21,6 +21,7 @@ pluginManagement {
 rootProject.name = "nx-cloud-ce"
 include(":apps:server")
 include(":libs:server:domain")
+include(":libs:server:gateway")
 include(":libs:server:persistence")
 include(":libs:server:storage:azure")
 include(":libs:server:storage:core")
